server/requirements.txt
django
djangorestframework
django-cors-headers
django[bcrypt]
Redis